+++ aiowsgi/compat.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+try:
+    import trollius as asyncio
+except ImportError:
+    import asyncio  # NOQA